scripts/pyproject.toml
[tool.poetry]
name = "pyspark scripts"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
invoke = "~1.7"
pyspark = "~3.3"
pandas = "~1.5"
pyarrow = "==10.0.1"
kafka-python = "~2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
